chore(registration-react): Move connected-feature-toggle-provider under directory feature-toggle
