+++ applications/registration-react/src/components/feature-toggle/connected-feature-toggle-provider.js
